requirements.txt
mistralai>=0.0.7
docling>=0.4.0
pymupdf4llm>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
